# Conflicts:
#	study/package-lock.json
#	study/package.json

+++ core/src/org/labkey/core/analytics/AnalyticsServiceImpl.java
@@ -162,7 +162,9 @@ public class AnalyticsServiceImpl implements AnalyticsService
     {
         ActionURL actionUrl = context.cloneActionURL();
         Container container = context.getContainer();
-        if (!container.hasPermission(UserManager.getGuestUser(), ReadPermission.class))
+
+        // Adding a null check for container as on rendering the error page, container can be null for a not found page
+        if (null != container && !container.hasPermission(UserManager.getGuestUser(), ReadPermission.class))
         {
             actionUrl.deleteParameters();
             actionUrl.setExtraPath(container.getId());
